Move positive_int to common_args
